Move capabilities model to apitype module
Keep interpretation of the response with the type itself instead of with the backend.
